#5 Added possibility to fetch reviews through reviews manager. Added tests and updated tests & packages structure a little.
