+++ lib/src/managers/locations_manager.dart
@@ -1,10 +1,9 @@
 import 'dart:convert' show json;
 
 import 'package:http/http.dart' as http;
-import 'package:http/http.dart';
 
-import 'gmb_api.dart';
-import 'models/location/location.dart';
+import '../gmb_api.dart';
+import '../models/location/location.dart';
 
 /// Responsible for retrieving and managing locations for the given account (id).
 class LocationsManager {
@@ -19,7 +18,7 @@ class LocationsManager {
   Future<void> fetchLocations(Function(List<Location> locations) onSuccess,
       Function(http.Response response) onError,
       [http.Client httpClient]) async {
-    if (httpClient == null) httpClient = Client();
+    if (httpClient == null) httpClient = http.Client();
 
     final http.Response response = await httpClient.get(
       'https://mybusiness.googleapis.com/v4/accounts/$accountId/locations',
